+++ src/main/javascript/Mailchimp/index.js
@@ -1,6 +1,0 @@
-
-export {FetchClient as MailchimpFetchClient} from './APIClient'
-export {fetchSubscriberInfo, fetchLists, updateListSubscriptions, determineMembershipDetails, determineAverageRating} from './Services'
-export {ApiCredentials, MailchimpAuthcInfo} from './Authentication'
-
-export {AuthenticationError as MailchimpAuthenticationError} from './Errors'